requirements.txt
sentence-transformers
chromadb
fastapi
uvicorn[standard]
numpy
pandas
scikit-learn
umap-learn
joblib